Created page and ViewModel for maintenance reminder
