+++ BFH_USZ_PICC/BFH_USZ_PICC/BFH_USZ_PICC.csproj
@@ -142,6 +142,7 @@
     <Compile Include="ViewModels\GlossaryViewModel.cs" />
     <Compile Include="ViewModels\KnowledgeEntriesViewModel.cs" />
     <Compile Include="ViewModels\FormerPICCDetailViewModel.cs" />
+    <Compile Include="ViewModels\MaintenanceReminderViewModel.cs" />
     <Compile Include="ViewModels\SettingsViewModel.cs" />
     <Compile Include="ViewModels\MasterDataViewModel.cs" />
     <Compile Include="ViewModels\MaintenanceInstructionViewModel.cs" />
@@ -192,6 +193,9 @@
     </Compile>
     <Compile Include="Views\FormerPICCDetailPage.xaml.cs">
       <DependentUpon>FormerPICCDetailPage.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="Views\MaintenanceReminderPage.xaml.cs">
+      <DependentUpon>MaintenanceReminderPage.xaml</DependentUpon>
     </Compile>
     <Compile Include="Views\UserMasterDataPage.xaml.cs">
       <DependentUpon>UserMasterDataPage.xaml</DependentUpon>
@@ -515,7 +519,12 @@
       <SubType>Designer</SubType>
     </EmbeddedResource>
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <EmbeddedResource Include="Views\MaintenanceReminderPage.xaml">
+      <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
+      <SubType>Designer</SubType>
+    </EmbeddedResource>
+  </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
   <Import Project="..\..\packages\Xamarin.Forms.2.3.3.175\build\portable-win+net45+wp80+win81+wpa81+MonoAndroid10+Xamarin.iOS10+xamarinmac20\Xamarin.Forms.targets" Condition="Exists('..\..\packages\Xamarin.Forms.2.3.3.175\build\portable-win+net45+wp80+win81+wpa81+MonoAndroid10+Xamarin.iOS10+xamarinmac20\Xamarin.Forms.targets')" />
   <Import Project="$(MSBuildExtensionsPath)\Microsoft\Multilingual App Toolkit\Microsoft.Multilingual.ResxResources.targets" Label="MultilingualAppToolkit" Condition="Exists('$(MSBuildExtensionsPath)\Microsoft\Multilingual App Toolkit\v$(MultilingualAppToolkitVersion)\Microsoft.Multilingual.ResxResources.targets')" />
