+++ BFH_USZ_PICC/BFH_USZ_PICC/MultilingualResources/BFH_USZ_PICC.de.xlf
