+++ BFH_USZ_PICC/BFH_USZ_PICC/ViewModels/ViewModelLocator.cs
@@ -62,6 +62,7 @@ namespace BFH_USZ_PICC.ViewModels
             SimpleIoc.Default.Register<MasterDataViewModel>();
             SimpleIoc.Default.Register<SettingsViewModel>();
             SimpleIoc.Default.Register<FormerPICCDetailViewModel>();
+            SimpleIoc.Default.Register<MaintenanceReminderViewModel>();     
             
         }
