+++ BFH_USZ_PICC/BFH_USZ_PICC/Views/MaintenanceReminderPage.xaml
@@ -1,0 +1,40 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<d:BaseContentPage xmlns="http://xamarin.com/schemas/2014/forms"
+                   xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+                   x:Class="BFH_USZ_PICC.Views.MaintenanceReminderPage"
+                   xmlns:utilitys="clr-namespace:BFH_USZ_PICC.Utilitys"
+                   xmlns:d="clr-namespace:BFH_USZ_PICC.Views;assembly=BFH_USZ_PICC.Views"
+                   xmlns:i18n="clr-namespace:BFH_USZ_PICC.Utilitys.Extensions"
+                   Title="{i18n:Translate SettingsPageMaintenanceReminderText}"
+                   BindingContext="{Binding [MaintenanceReminderViewModel], Source={StaticResource ViewModelLocator}}">
+  <ScrollView>
+    <StackLayout Orientation="Vertical"
+                 VerticalOptions="FillAndExpand"
+                 Margin="15">
+      <Label Text="Das ist ein Text"/>
+      <Label Text="{i18n:Translate SettingsPageMaintenanceReminderText}"/>
+      <Switch IsToggled="{Binding IsReminderSet, Mode=TwoWay}"
+              HorizontalOptions = "Start"/>
+      <StackLayout x:Name="EditWeeklyReminder"
+                   IsVisible="{Binding IsReminderSet}">
+        <Label Text="{i18n:Translate SettingsPageMaintenanceReminderSetStartDateText}"/>
+        <DatePicker Date="{Binding ReminderStartDate}"
+                    IsEnabled="{Binding IsReminderEditable}"/>
+        <Label Text="{i18n:Translate SettingsPageMaintenanceReminderSetDailyTimeText}"/>
+        <TimePicker Time="{Binding ReminderDayTime}"
+                    IsEnabled="{Binding IsReminderEditable}"/>
+        <Label Text="{i18n:Translate SettingsPageMaintenanceReminderFrequencyText}"/>
+        <Picker x:Name="Frequency"
+                SelectedIndex="{Binding ReminderFrequency}"
+                IsEnabled="{Binding IsReminderEditable}"/>
+        <Label Text="{i18n:Translate SettingsPageMaintenanceReminderRepetitionText}"/>
+        <Picker x:Name="Repetition"
+                SelectedIndex="{Binding ReminderRepetition}"
+                IsEnabled="{Binding IsReminderEditable}"/>
+        <Button Text="{i18n:Translate SettingsPageAddReminderButtonText}"
+                Command="{Binding AddNotificationsCommand}"
+                IsEnabled="{Binding IsReminderEditable}"/>
+      </StackLayout>
+    </StackLayout>
+  </ScrollView>
+</d:BaseContentPage>
